FormController. Обработка ошибок.

+++ src/main/java/com/test/task/controllers/FormController.java
@@ -2,7 +2,11 @@ package com.test.task.controllers;
 
 import com.test.task.entities.Form;
 import com.test.task.entities.dtos.FormDto;
+import com.test.task.exceptions.MalformedEntityException;
+import com.test.task.exceptions.NullIdException;
+import com.test.task.exceptions.nonExistentIdException;
 import com.test.task.services.FormService;
+import com.test.task.utils.StringConstants;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -31,19 +35,24 @@ public class FormController {
     public ResponseEntity<?> addForm(@RequestBody @Validated Form form, BindingResult bindingResult){
         if(form.getId() == null)
             form.setId(null);
+        if (bindingResult.hasErrors())
+            throw new MalformedEntityException();
        return new ResponseEntity<>(formService.saveOrUpdate(form), HttpStatus.OK);
     }
 
     @PutMapping
     public ResponseEntity<?> updateForm(@RequestBody @Validated Form form, BindingResult bindingResult){
         if(form.getId() == null)
-            throw new RuntimeException("id can't be null");
+            throw new NullIdException();
+        if (bindingResult.hasErrors())
+            throw new MalformedEntityException();
         return new ResponseEntity<>(formService.saveOrUpdate(form), HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> delete(@PathVariable Long id){
+    public void delete(@PathVariable Long id){
+        if (!formService.existsById(id))
+            throw new nonExistentIdException(StringConstants.noObjectWithThisId);
         formService.deleteById(id);
-        return new ResponseEntity<>(HttpStatus.OK);
     }
 }
